+++ packages/components/xpub-dashboard/src/components/Dashboard.js
@@ -13,6 +13,7 @@ const EditorItemWithVersion = withVersion(EditorItem)
 const ReviewerItemWithVersion = withVersion(ReviewerItem)
 
 const Dashboard = ({
+  acceptFiles,
   currentUser,
   conversion,
   dashboard,
@@ -23,6 +24,7 @@ const Dashboard = ({
   <Page>
     <UploadContainer>
       <UploadManuscript
+        acceptFiles={acceptFiles}
         conversion={conversion}
         uploadManuscript={uploadManuscript}
       />
@@ -38,7 +40,7 @@ const Dashboard = ({
       object={dashboard}
       operation="can view my submission section"
     >
-      {dashboard.length > 0 && (
+      {dashboard.length > 0 ? (
         <Section>
           <Heading>My Submissions</Heading>
           {dashboard.map(project => (
@@ -54,11 +56,11 @@ const Dashboard = ({
             />
           ))}
         </Section>
-      )}
+      ) : null}
     </Authorize>
 
     <Authorize object={dashboard} operation="can view review section">
-      {dashboard.length > 0 && (
+      {dashboard.length > 0 ? (
         <Section>
           <Heading>To review</Heading>
           {dashboard.map(project => (
@@ -70,7 +72,7 @@ const Dashboard = ({
             />
           ))}
         </Section>
-      )}
+      ) : null}
     </Authorize>
 
     <Authorize
@@ -78,7 +80,7 @@ const Dashboard = ({
       object={dashboard}
       operation="can view my manuscripts section"
     >
-      {dashboard.length > 0 && (
+      {dashboard.length > 0 ? (
         <Section>
           <Heading>My Manuscripts</Heading>
           {dashboard.map(project => (
@@ -88,7 +90,7 @@ const Dashboard = ({
             />
           ))}
         </Section>
-      )}
+      ) : null}
     </Authorize>
   </Page>
 )