+++ packages/components/xpub-review/src/components/review/ReviewLayout.js
@@ -1,7 +1,6 @@
 import React from 'react'
 
 import moment from 'moment'
-// import classnames from 'classnames'
 import { Tabs } from '@pubsweet/ui'
 import SimpleEditor from 'wax-editor-react'
 
@@ -9,6 +8,7 @@ import ReviewForm from './ReviewForm'
 import ReviewMetadata from '../metadata/ReviewMetadata'
 import Review from './Review'
 import { Columns, Manuscript, Admin } from '../atoms/Columns'
+import { Info } from '../molecules/Info'
 
 const ReviewLayout = ({
   project,
@@ -52,15 +52,19 @@ const ReviewLayout = ({
 
       // TODO: need to include unreviewed versions?
       editorSections.push({
-        content: (
-          <SimpleEditor
-            content={version.source}
-            editing="selection"
-            key={key}
-            layout="bare"
-            readOnly
-          />
-        ),
+        content:
+          lastSubmitted.files.manuscript.type ===
+          'application/vnd.openxmlformats-officedocument.wordprocessingml.document' ? (
+            <SimpleEditor
+              content={lastSubmitted.source}
+              editing="selection"
+              key={key}
+              layout="bare"
+              readOnly
+            />
+          ) : (
+            <Info>No supported view of the file</Info>
+          ),
         key,
         label,
       })
@@ -96,15 +100,19 @@ const ReviewLayout = ({
     })
 
     editorSections.push({
-      content: (
-        <SimpleEditor
-          content={lastSubmitted.source}
-          editing="selection"
-          key={key}
-          layout="bare"
-          readOnly
-        />
-      ),
+      content:
+        lastSubmitted.files.manuscript.type ===
+        'application/vnd.openxmlformats-officedocument.wordprocessingml.document' ? (
+          <SimpleEditor
+            content={lastSubmitted.source}
+            editing="selection"
+            key={key}
+            layout="bare"
+            readOnly
+          />
+        ) : (
+          <Info>No supported view of the file</Info>
+        ),
       key,
       label,
     })