+++ packages/server/src/start-server.js
@@ -3,6 +3,8 @@ const config = require('config')
 const Promise = require('bluebird')
 const logger = require('@pubsweet/logger')
 
+const { addSubscriptions } = require('./graphql/subscriptions')
+
 const startServer = async app => {
   const port =
     config['pubsweet-server'].port ||
@@ -14,6 +16,7 @@ const startServer = async app => {
   const startListening = Promise.promisify(server.listen, { context: server })
   await startListening(port)
   logger.info(`App is listening on port ${port}`)
+  addSubscriptions(server)
   return server
 }