+++ packages/components/TeamsManager/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="1.1.17"></a>
+## [1.1.17](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-teams-manager@1.1.16...pubsweet-component-teams-manager@1.1.17) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-teams-manager
+
+<a name="1.1.16"></a>
+## [1.1.16](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-teams-manager@1.1.15...pubsweet-component-teams-manager@1.1.16) (2018-07-27)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-teams-manager
+
 <a name="1.1.15"></a>
 ## [1.1.15](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-teams-manager@1.1.14...pubsweet-component-teams-manager@1.1.15) (2018-07-19)