+++ packages/components/xpub-review/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="3.1.5"></a>
+## [3.1.5](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-review@3.1.4...pubsweet-component-xpub-review@3.1.5) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-xpub-review
+
+<a name="3.1.4"></a>
+## [3.1.4](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-review@3.1.3...pubsweet-component-xpub-review@3.1.4) (2018-07-27)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-xpub-review
+
 <a name="3.1.3"></a>
 ## [3.1.3](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-review@3.1.2...pubsweet-component-xpub-review@3.1.3) (2018-07-23)