+++ packages/components/xpub-submit/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+<a name="4.0.7"></a>
+## [4.0.7](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-submit@4.0.6...pubsweet-component-xpub-submit@4.0.7) (2018-08-02)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-xpub-submit
+
+<a name="4.0.6"></a>
+## [4.0.6](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-submit@4.0.5...pubsweet-component-xpub-submit@4.0.6) (2018-07-27)
+
+
+
+
+**Note:** Version bump only for package pubsweet-component-xpub-submit
+
 <a name="4.0.5"></a>
 ## [4.0.5](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/pubsweet-component-xpub-submit@4.0.4...pubsweet-component-xpub-submit@4.0.5) (2018-07-23)