+++ packages/ui/CHANGELOG.md
@@ -3,7 +3,29 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
-  <a name="8.1.0"></a>
+  <a name="8.3.0"></a>
+# [8.3.0](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui@8.2.0...@pubsweet/ui@8.3.0) (2018-08-02)
+
+
+### Features
+
+* **ui-toolkit:** add borderColor function to theme helper ([0ed3fc1](https://gitlab.coko.foundation/pubsweet/pubsweet/commit/0ed3fc1))
+
+
+
+
+  <a name="8.2.0"></a>
+# [8.2.0](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui@8.1.0...@pubsweet/ui@8.2.0) (2018-07-27)
+
+
+### Features
+
+* add Attachments pubsweet comp for image upload ([a2dc8ca](https://gitlab.coko.foundation/pubsweet/pubsweet/commit/a2dc8ca))
+
+
+
+
+<a name="8.1.0"></a>
 # [8.1.0](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui@8.0.1...@pubsweet/ui@8.1.0) (2018-07-12)
 
 
@@ -20,7 +42,7 @@ See [Conventional Commits](https://conventionalcommits.org) for commit guideline
 
 
 
-  <a name="8.0.1"></a>
+<a name="8.0.1"></a>
 ## [8.0.1](https://gitlab.coko.foundation/pubsweet/pubsweet/compare/@pubsweet/ui@8.0.0...@pubsweet/ui@8.0.1) (2018-07-09)